first batch of migrating old-tck to new-tck; the simplest ones: ViewHandler/ProtectedViews
