+++ tck/faces20/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2023 Contributors to Eclipse Foundation.
+    Copyright (c) 2026 Contributors to Eclipse Foundation.
 
     This program and the accompanying materials are made available under the
     terms of the Eclipse Public License v. 2.0, which is available at
@@ -25,15 +25,15 @@
         <version>5.0.0-SNAPSHOT</version>
     </parent>
 
-    <groupId>org.eclipse.ee4j.tck.faces.old-tck-selenium</groupId>
+    <groupId>org.eclipse.ee4j.tck.faces.faces20</groupId>
     <artifactId>pom</artifactId>
     <packaging>pom</packaging>
 
-    <name>Jakarta Faces TCK ${project.version} - Test - Old TCK Selenium</name>
+    <name>Jakarta Faces TCK ${project.version} - Test - Faces 2.0</name>
 
     <modules>
-        <module>ajax</module>
-        <module>commandLink</module>
+        <module>protectedView</module>
+        <module>viewhandler</module>
     </modules>
 
     <dependencies>
@@ -41,6 +41,7 @@
             <groupId>org.eclipse.ee4j.tck.faces.test</groupId>
             <artifactId>util</artifactId>
             <version>${project.version}</version>
+            <scope>test</scope>
         </dependency>
     </dependencies>
 </project>
