+++ tck/faces20/protectedView/src/main/webapp/WEB-INF/faces-config.xml
