--- tck/old-tck-selenium/protectedViews/pom.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2023 Contributors to Eclipse Foundation.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0, which is available at
-    http://www.eclipse.org/legal/epl-2.0.
-
-    This Source Code may also be made available under the following Secondary
-    Licenses when the conditions for such availability set forth in the
-    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
-    version 2 with the GNU Classpath Exception, which is available at
-    https://www.gnu.org/software/classpath/license.html.
-
-    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-
--->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.eclipse.ee4j.tck.faces.old-tck-selenium</groupId>
-        <artifactId>pom</artifactId>
-        <version>5.0.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>protectedViews</artifactId>
-    <packaging>war</packaging>
-
-    <name>Jakarta Faces TCK ${project.version} - Test - Faces 2.2 - protectedViews</name>
-
-    <build>
-        <finalName>old-tck-protectedViews</finalName>
-    </build>
-</project>
