--- tck/old-tck-selenium/protectedViews/src/main/webapp/views/protected.xhtml
@@ -1,31 +1,0 @@
-<!--
-
-    Copyright (c) 2012, 2022 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0, which is available at
-    http://www.eclipse.org/legal/epl-2.0.
-
-    This Source Code may also be made available under the following Secondary
-    Licenses when the conditions for such availability set forth in the
-    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
-    version 2 with the GNU Classpath Exception, which is available at
-    https://www.gnu.org/software/classpath/license.html.
-
-    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-
--->
-
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en"
-      xmlns:h="jakarta.faces.html"
-      xmlns:f="jakarta.faces.core"> 
-	
-	<h:head>
-		<title>Protected View</title>
-	</h:head> 
-	<h:body>
-		<h:outputText id="messOne" value="This is a Protected View!" />
-	</h:body> 
-	
-</html>
