--- tck/old-tck-selenium/protectedViews/src/main/webapp/views/public.xhtml
@@ -1,31 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<!--
-
-    Copyright (c) 2013, 2022 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0, which is available at
-    http://www.eclipse.org/legal/epl-2.0.
-
-    This Source Code may also be made available under the following Secondary
-    Licenses when the conditions for such availability set forth in the
-    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
-    version 2 with the GNU Classpath Exception, which is available at
-    https://www.gnu.org/software/classpath/license.html.
-
-    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-
--->
-
-<xhtml xmlns="http://www.w3.org/1999/xhtml"
-	xmlns:h="jakarta.faces.html"
-	xmlns:f="jakarta.faces.core"> <h:head>
-	<title>Public View</title>
-</h:head> <h:body>
-	<h:outputText value="This is a Public View!" />
-	<h:form id="form1">
-		<h:commandLink id="linkOne" value="linkOne" action="protected.xhtml">
-			<h:outputText value="To Protected View" />
-		</h:commandLink>
-	</h:form>
-</h:body> </xhtml>
