+++ tck/old-tck/run/src/test/etc/ts-all-platform.jtx
@@ -5397,8 +5397,6 @@ com/sun/ts/tests/jsf/spec/templating/remove/URLClient.java#templateUIRemoveTagTe
 com/sun/ts/tests/jsf/spec/templating/repeat/URLClient.java#templateUIRepeatOffsetTest
 com/sun/ts/tests/jsf/spec/templating/repeat/URLClient.java#templateUIRepeatVarStatusTest
 com/sun/ts/tests/jsf/spec/templating/repeat/URLClient.java#templateUIRepeatVarTest
-com/sun/ts/tests/jsf/spec/view/viewhandler/URLClient.java#viewHandlerCreateViewTest
-com/sun/ts/tests/jsf/spec/view/viewhandler/URLClient.java#viewHandlerRestoreViewTest
 com/sun/ts/tests/jsf/spec/webapp/factoryfinder/URLClient.java#factoryFinderConfig1Test
 com/sun/ts/tests/jsf/spec/webapp/factoryfinder/URLClient.java#factoryFinderConfig2Test
 com/sun/ts/tests/jsf/spec/webapp/factoryfinder/URLClient.java#factoryFinderConfig3Test
