--- tck/old-tck/source/src/com/sun/ts/tests/jsf/spec/view/viewhandler/build.xml
@@ -1,45 +1,0 @@
-<!--
-
-    Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0, which is available at
-    http://www.eclipse.org/legal/epl-2.0.
-
-    This Source Code may also be made available under the following Secondary
-    Licenses when the conditions for such availability set forth in the
-    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
-    version 2 with the GNU Classpath Exception, which is available at
-    https://www.gnu.org/software/classpath/license.html.
-
-    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-
--->
-
-<project name="jsf_view_viewhandler" basedir="." default="usage">
-    
-    <!-- IMPORTS -->
-    <import file="../../../common/common.xml"/>
-    
-    <property name="app.name" value="jsf_view_viewhandler"/>
-    <property name="facelets.dir" 
-              value="${ts.home}/src/web/jsf/spec/view/viewhandler"/>
-    
-    <target name="package">
-        <ts.war archivename="${app.name}"
-                descriptor="${app.name}_web.xml">
-            <zipfileset dir="${class.dir}"
-                        includes="${jsf.common}/servlets/HttpTCKServlet.class,
-                        ${jsf.common}/util/JSFTestUtil.class"
-                        prefix="WEB-INF/classes"/>
-            <zipfileset dir="WEB-INF" includes="*.xml"
-                        prefix="WEB-INF"/>
-            <fileset dir="${facelets.dir}" includes="*.xhtml"/>
-            <zipfileset dir="${ts.home}/weblib"
-                        includes="*.jar"
-                        prefix="WEB-INF/lib"/>
-            <zipfileset dir="${src.dir}/${jsf.common}" includes="beans.xml" prefix="WEB-INF"/>
-        </ts.war>      
-    </target>
-    
-</project>
